requirements.txt
speechrecognition
pyttsx3
pywhatkit
opencv-python
pillow
face-recognition

